Fix System.Http API usage
Before, we were trying to call GetHeader on a HttpResponseMessage object. This
would not return any content-* headers - those are accessed through the
Content.Headers object.

Additionally, the integration tests were hard to run, so this change also adds
the ability to pass a flag "integration" to test.sh so you can run them easily.