Add .NET Framework 4.5 and .NET Standard 2.0 as framework targets
This commit changes this project to target both .NET Framework 4.5 and
.NET Standard 2.0 where originally we only targeted .NET Framework 4.5.
.NET Standard 2.0 allows this client to be used in .NET Core
applications which is an alternative .NET framework specifically made
for cloud applications.

No code changes were necessary for the library itself, only test code
needed to be changed. There's only the addition of one preprocessor
directive in test code.

Additionally, I've updated these projects to use the new .csproj/.vbproj
format:

- Phaxio
- Phaxio.Examples.Basic
- Phaxio.Examples.Basic.VbNet
- Phaxio.Test

Phaxio.Examples.ReceiveCallback was not changed since the new .csproj
format does not work for ASP.NET projects.

The test script (test.sh) was also updated to test both systems and to
switch to msbuild (since xbuild is being deprecated and Travis CI will
soon switch to msbuild). The test script was also updated so integration
tests could be run if you call it like so "./test.sh integration".

The Travis CI configuration also needed to be updated to specifically
build and test both frameworks, and we needed to make sure that the new
.NET Core SDK was installed. We're also now specifying which version of
Mono we're building the .NET Framework 4.5 version against.

Finally, this updates the documentation with the new requirements and
commands necessary to build, test, and package this project.