Update library to use Phaxio API to use v2.1
Phaxio has released a new API, so we need to update this client library to
support those changes. There are three major updates: a new endpoint,
removing parameters from the fax callback, and moving to HTTP Basic
Authorization.

1) We're going to use a new endpoint for the library (a different version
string), which was just changing a string in the `PhaxioClient` class.

2) The callback example project needed to remove `success`, `is_test`,
and `direction` from the expected parameters.

3) We needed to move to HTTP Basic Authorization. I added a
BasicAuthorization object that contained & handled this change, and then
updated IRestRequest and RestRequest to have a field called Authorization
that's a BasicAuthorizationType. The RestClient now checks to see if the
IRestRequest has this field and if so, sets the HttpClient's
DefaultRequestHeaders.Authorization. Finally, the BasePhaxioClient object
now adds this object to the IRestRequest instead of setting the key and
secret as request parameters.

Finally, `Barcodes` (an array of Barcode objects) and `CallerName` (a
string) were added the Fax object.