Fix 500 errors with sample callback project
This project needed several updates to be able to run without giving an error.
This modifies the webhook URL to actually get the callback form data, which
it was not getting. This was done by using async/await vs. using the ContinueWith
method on Task.

The trigger controller was changed to MVC vs. WebAPI since it was being used for
a regular HTML form. This allowed us to just redirect vs. showing an XML page.

I also updated the index to have a few more instructions.